requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
pyyaml
python-dotenv
nltk
rouge-score